Port subFiller and combine with addFiller.
